subscribe to event when loading plugin
