+++ ConanToolWindowControl.xaml.cs
@@ -14,6 +14,7 @@ using EnvDTE;
 using Microsoft.VisualStudio.Threading;
 using System.Windows.Navigation;
 using System.Windows.Media;
+using Microsoft.VisualStudio.PlatformUI;
 
 namespace conan_vs_extension
 {
@@ -58,8 +59,32 @@ namespace conan_vs_extension
             this.InitializeComponent();
             LibraryHeader.Visibility = Visibility.Collapsed;
 
+            this.Loaded += ConanToolWindowControl_Loaded;
+
             ToggleUIEnableState(IsConanInitialized());
             _ = InitializeAsync();
+        }
+
+        private void ConanToolWindowControl_Loaded(object sender, RoutedEventArgs e)
+        {
+            VSColorTheme.ThemeChanged += OnThemeChanged;
+            UpdateTheme();
+        }
+
+        private void OnThemeChanged(ThemeChangedEventArgs e)
+        {
+            ThreadHelper.ThrowIfNotOnUIThread();
+            UpdateTheme();
+        }
+
+        public void UpdateTheme()
+        {
+            ThreadHelper.ThrowIfNotOnUIThread();
+
+            var currentThemeColor = VSColorTheme.GetThemedColor(EnvironmentColors.ToolWindowTextColorKey);
+            var currentColor = Color.FromRgb(currentThemeColor.R, currentThemeColor.G, currentThemeColor.B);
+
+            UpdateForeground(currentColor);
         }
 
         public void UpdateForeground(Color color)
